Update HypixelGuild.java
This commit will add more skeleton methods to the HypixelGuild class.

The Hypixel API does not support leaderboard placement retrieval for guilds. The Sk1er and plancke API's both support the retrieval of that data. In order to retrieve the leaderboard placement of a Guild it may be necessary to use them.