Delete HypixelAPIDataRetrieval.java
This commit will delete an unneeded class. The utility of this class was that it could cast the JSON object to an Object of the developers choice. This is bad because:

A. The casting would not work because you can’t just cast from a String to any given type of Object.

B. The getInformation() method is already present in the RequestFactory class. That method returns a String allowing the caller to cast the String to whatever type of object they want.