Create all of the modes of BedWars
This commit will create all the different modes of BedWars as classes. It will then add all of the classes as dependency's to the `HypixelBedWars` class.

The Player class also has a method named getBedWarsStats(). It will return a `HypixelBedWars` Object. The `HypixelBedWars` Object will then have methods to reference all of it's dependency's.