Added the `prefix` field to each BedWars class
This commit will give each `HypixelBedWars` subclass a `prefix`. This will allow the superclass to use the prefix when retrieving information from the `JSONObject`.

If a subclass of `HypixelBedWars` has a different key for a field it can simply override the superclass method and pass in it's unique key.

These classes will be deleted because I can not find any way to retrieve information about them from the API:
-`Ultimate Solos`
-`Capture`
-`Lucky Fours`
-`Rush Solo`
-`Swap Doubles`
-`Swap Fours`
-`Ultimate Solos`

 They can be added if someone finds a way to retrieve data about them in subsequent commits.