Update HypixelPlayer.java
This commit will implement many of the methods in the `HypixelPlayer` class. It uses the `RequestFactory` to instantiate JSON Objects and retrieve information from them.

This commit will also add two methods:
- isLevelingRewardClaimed
It takes in a integer and retrieves a boolean stating if the users leveling reward of that level has been claimed.

-getUserLanguage
returns a String of the users selected language.