Implemented JSON into all methods in `HypixelBedWarsStats`
This commit will implement the JSON retrieval into many of the methods in the `HypixelBedWarsStats` class. It excludes a few methods. This is because an enum prototype is being constructed to retrieve those methods as an enum instead of a String. They will be left alone until that branch is done.