Forced send() to accept an `Endpoint` instead of a String
This commit will force the send() to accept an `Endpoint` instead of a String. This insures that invalid `URL`'s will not be passed in.

Also, a minor error that I noticed was that in the getInformation() and getEndpointThroughAPI() they converted the endpoints to `URL`'s by using toString() instead of getURL().