Added a `mode` field to `HypixelBedWars`
This commit will allow the API user to check the mode of the `HypixelBedWars` object they're using. The `mode` field is final, and this is no setter because it is meant to be immutable. In order to get information from a sub-mode a new `HypixelBedWars` would have to be created.