Implement getHypixelFriends() methods
This commit will implement the getHypixelFriends() methods. Essentially, it iterates through every `JSONObject` in the `records` `JSONArray` returned by the `friends Endpoint` and creates a `HypixelFriend` with the `UUID` and `Unix ID`.

This commit will also fix the getUUID() method in the `HypixelPlayer` class. It makes it return a `UUID` instead of a `String`.