Change JSONHandler from returning empty optionals to default values
This commit will make the `JSONHandler` class return default values instead of empty optionals. This is because it's much easier for the end user to deal with default values. For example, if they are trying to get a `HypixelPlayer`'s  bedwars experience, and no value is returned it just returns 0.